[update] 文字数と同じ名前が存在しないか確認するように変更

+++ src/pages/Signup.tsx
@@ -17,22 +17,34 @@ interface ResposBody {
 const Signup: React.FC = () => {
     const [name, setName] = React.useState<string>("");
     const [password, setPassword] = React.useState<string>("");
-    const navigate = useNavigate()
+    const [status, setStatus] = React.useState<string>("名前とパスワードを入力してね");
+    const navigate = useNavigate();
     const body: User = {
         name: name,
         password: password
     }
 
     // ユーザーを作成してユーザーIDを保存
-    const submit_user = async () =>  { 
+    const submit_user = async () =>  {
+        if (password.length <= 8) {
+            setStatus("パスワードは8桁以上にしてください");
+            return 
+        } 
         const {data}: AxiosResponse<ResposBody> = await axios.post(`${ROOT_URL}/regist`, body);
-        await localStorage.setItem("token", data.token)
-        await navigate("/", {state: {s:true}})
+        if (data.token == "exist name") {
+            setStatus("存在する名前なので名前を変えてください");
+        } else {
+            await localStorage.setItem("token", data.token);
+            await navigate("/", { state: { s: true } });
+        }
     }
 
     return (
         <div className="signup">
             <h2>サインアップ</h2>
+            <div>
+                <p>{status}</p>
+            </div>
             <div className="name-input">
                 <label>なまえ<input type="text" value={name} onChange={(event) => setName(event.target.value)}></input></label>
             </div>
